+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - (Feature) Windows Platform CLI
 - (Feature) (Platform) Auth User Creation
 - (Maintenance) Add Common Api Import
+- (Feature) Previous Pod Logs in DebugPackage
 
 ## [1.2.48](https://github.com/arangodb/kube-arangodb/tree/1.2.48) (2025-05-08)
 - (Maintenance) Extend Documentation